refactor: 도메인 패키지 생성 후 Lotto와 Money 클래스 이동
